refactor: 불필요한 코드 제거, 파일명 변경
