+++ src/main/java/re/kr/icuh/icuhplatform/controller/FileStorageController.java
@@ -5,7 +5,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
-import re.kr.icuh.icuhplatform.service.AttachmentService;
+import re.kr.icuh.icuhplatform.service.FileStorageService;
 
 import java.io.IOException;
 import java.util.List;
@@ -14,9 +14,9 @@ import java.util.List;
 @RestController
 @RequestMapping("/api/v1")
 @RequiredArgsConstructor
-public class AttachmentApiController {
+public class FileStorageController {
 
-    private final AttachmentService attachmentService;
+    private final FileStorageService fileStorageService;
 
     @GetMapping("/test")
     public String test() {
@@ -26,7 +26,7 @@ public class AttachmentApiController {
     @PostMapping("/attachments")
     public String attachments(@RequestPart("attachment") List<MultipartFile> file) throws IOException {
 
-        attachmentService.createAttachment(file);
+        fileStorageService.createAttachment(file);
 
         return "ok";
     }
@@ -34,7 +34,7 @@ public class AttachmentApiController {
     @PostMapping("/large-attachment")
     public ResponseEntity<String> largeAttachments(@RequestParam("attachment") MultipartFile multipartFile) throws Exception {
 
-        attachmentService.uploadLargeFile(multipartFile);
+        fileStorageService.uploadLargeFile(multipartFile);
 
         return ResponseEntity.ok("파일 업로드 성공");
     }
