+++ src/main/java/re/kr/icuh/icuhplatform/service/FileStorageService.java
@@ -5,12 +5,11 @@ import lombok.extern.slf4j.Slf4j;
 import org.apache.tomcat.util.http.fileupload.FileUploadException;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
-import re.kr.icuh.icuhplatform.common.FileMetadata;
-import re.kr.icuh.icuhplatform.common.FileUtils;
+import re.kr.icuh.icuhplatform.common.util.FileUtils;
 import re.kr.icuh.icuhplatform.domain.Attachment;
+import re.kr.icuh.icuhplatform.domain.FileMetadata;
 import re.kr.icuh.icuhplatform.dto.CreateAttachmentDto;
-import re.kr.icuh.icuhplatform.repository.AttachmentRepository;
-import re.kr.icuh.icuhplatform.util.S3FileUploader;
+import re.kr.icuh.icuhplatform.repository.FileStorageRepository;
 
 import java.io.File;
 import java.io.IOException;
@@ -19,11 +18,11 @@ import java.util.List;
 @Slf4j
 @Service
 @RequiredArgsConstructor
-public class AttachmentService {
+public class FileStorageService {
 
     private final FileUtils fileUtils;
     private final S3FileUploader s3FileUploader;
-    private final AttachmentRepository attachmentRepository;
+    private final FileStorageRepository fileStorageRepository;
 
     public void createAttachment(List<MultipartFile> file) throws IOException {
 
@@ -31,7 +30,7 @@ public class AttachmentService {
 
         for (CreateAttachmentDto attachmentDto : attachmentDtos) {
             Attachment attachment = attachmentDto.toAttachment(attachmentDto);
-            attachmentRepository.save(attachment);
+            fileStorageRepository.save(attachment);
         }
     }
 
@@ -53,7 +52,7 @@ public class AttachmentService {
                     .build();
 
             Attachment attachment = dto.toAttachment(dto);
-            attachmentRepository.save(attachment);
+            fileStorageRepository.save(attachment);
 
         } catch (Exception e) {
             log.error("파일 업로드 실패", e);
