+++ src/test/java/re/kr/icuh/icuhplatform/util/S3FileUploaderTest.java
@@ -11,8 +11,9 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.mock.web.MockMultipartFile;
 import org.springframework.test.util.ReflectionTestUtils;
 import org.springframework.web.multipart.MultipartFile;
-import re.kr.icuh.icuhplatform.common.FileUtils;
+import re.kr.icuh.icuhplatform.common.util.FileUtils;
 import re.kr.icuh.icuhplatform.dto.CreateAttachmentDto;
+import re.kr.icuh.icuhplatform.service.S3FileUploader;
 
 import java.io.IOException;
 import java.net.URL;
